jobs_parlour_backend/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django>=5.0,<6.0
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
gunicorn==23.0.0
